refactor: add mercator util for parsing
